+++ v2/behaviors/broadcaster_test.go
@@ -14,10 +14,10 @@ package behaviors_test
 import (
 	"testing"
 
-	"git.tideland.biz/gocn/behaviors"
-	"git.tideland.biz/gocn/cells"
-	"git.tideland.biz/gocn/testsupport"
-	"git.tideland.biz/gots/asserts"
+	"github.com/tideland/gocn/v2/behaviors"
+	"github.com/tideland/gocn/v2/cells"
+	"github.com/tideland/gocn/v2/testsupport"
+	"github.com/tideland/gots/v3/asserts"
 )
 
 //--------------------