+++ v2/behaviors/collector_test.go
@@ -14,9 +14,10 @@ package behaviors_test
 import (
 	"testing"
 
-	"git.tideland.biz/gocn/behaviors"
-	"git.tideland.biz/gocn/cells"
-	"git.tideland.biz/gots/asserts"
+	"github.com/tideland/gocn/v2/behaviors"
+	"github.com/tideland/gocn/v2/cells"
+	"github.com/tideland/gocn/v2/testsupport"
+	"github.com/tideland/gots/v3/asserts"
 )
 
 //--------------------
@@ -35,6 +36,9 @@ func TestCollectorBehavior(t *testing.T) {
 	for i := 0; i < 25; i++ {
 		env.Raise("collector", "collect", i)
 	}
+
+	testsupport.LetItWork()
+	testsupport.LetItWork()
 
 	var collected []behaviors.EventData
 	err := env.Request("collector", "collected", nil, &collected)