+++ v2/behaviors/errors.go
@@ -12,8 +12,8 @@ package behaviors
 //--------------------
 
 import (
-	"git.tideland.biz/goas/errors"
-	"git.tideland.biz/gocn/cells"
+	"github.com/tideland/goas/v3/errors"
+	"github.com/tideland/gocn/v2/cells"
 )
 
 //--------------------
@@ -21,15 +21,16 @@ import (
 //--------------------
 
 const (
-	ecNoError = iota
-	ecIllegalRequest
-	ecIllegalResponse
-	ecCannotRecover
-
-	msgIllegalRequest  = "cell does not understand request %q"
-	msgIllegalResponse = "response has illegal type: %v"
-	msgCannotRecover   = "cannot recover cell: %v"
+	ErrIllegalRequest = iota + 1
+	ErrIllegalResponse
+	ErrCannotRecover
 )
+
+var errorMessages = map[int]string{
+	ErrIllegalRequest:  "cell does not understand request %q",
+	ErrIllegalResponse: "response has illegal type: %v",
+	ErrCannotRecover:   "cannot recover cell: %v",
+}
 
 //--------------------
 // ERRORS
@@ -37,29 +38,29 @@ const (
 
 // NewIllegalRequestError returns an error describing an illegal request.
 func NewIllegalRequestError(request cells.Event) error {
-	return errors.New(ecIllegalRequest, msgIllegalRequest, request.Topic())
+	return errors.New(ErrIllegalRequest, errorMessages, request.Topic())
 }
 
 // IsIllegalRequestError checks if an error shows an illegal request.
 func IsIllegalRequestError(err error) bool {
-	return errors.IsError(err, ecIllegalRequest)
+	return errors.IsError(err, ErrIllegalRequest)
 }
 
 // IsIllegalResponseError checks if an error shows an illegal response.
 func IsIllegalResponseError(err error) bool {
-	return errors.IsError(err, ecIllegalResponse)
+	return errors.IsError(err, ErrIllegalResponse)
 }
 
 // NewCannotRecoverError returns an error showing that a cell cannot
 // recover from errors.
 func NewCannotRecoverError(err interface{}) error {
-	return errors.New(ecCannotRecover, msgCannotRecover, err)
+	return errors.New(ErrCannotRecover, errorMessages, err)
 }
 
 // IsCannotRecoverError checks if an error shows a cell that cannot
 // recover.
 func IsCannotRecoverError(err error) bool {
-	return errors.IsError(err, ecCannotRecover)
+	return errors.IsError(err, ErrCannotRecover)
 }
 
 // EOF