+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Tideland Go Cell Network
 ========================
 
+2014-04-20
+----------
+
+- Adopted changes of goas/errors
+
 2014-04-17
 ----------